+++ spec/spec_helper.rb
@@ -6,6 +6,8 @@ require "graphql"
 require "graphql-hive"
 require "vcr"
 require "webmock/rspec"
+require "super_diff/rspec"
+require "timecop"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure